@@ -12,24 +12,28 @@
     "@fortawesome/fontawesome-svg-core": "6.1.0",
     "@fortawesome/free-solid-svg-icons": "6.1.0",
     "@fortawesome/react-fontawesome": "0.1.18",
-    "@multiversx/sdk-dapp": "2.14.4",
+    "@multiversx/sdk-dapp": "2.14.6",
+    "@multiversx/sdk-core": "11.4.1",
+    "@multiversx/sdk-network-providers": "1.3.0",
     "@next/font": "13.1.1",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
     "axios": "0.24.0",
     "bootstrap": "4.6.2",
-    "eslint": "8.35.0",
-    "eslint-config-next": "13.2.3",
-    "moment": "2.29.1",
+    "autoprefixer": "10.4.14",
+    "eslint": "8.42.0",
+    "eslint-config-next": "13.4.4",
+    "moment": "2.29.4",
     "next": "13.4.4",
+    "next-transpile-modules": "^10.0.0",
     "react": "18.2.0",
     "react-bootstrap": "2.1.2",
     "react-dom": "18.2.0",
     "sass": "1.57.1",
-    "typescript": "4.9.5"
+    "typescript": "5.1.3"
   },
   "devDependencies": {
+    "@types/node": "20.2.5",
+    "@types/react": "18.2.8",
+    "@types/react-dom": "18.2.4",
     "@typescript-eslint/eslint-plugin": "^5.54.1",
     "eslint-config-prettier": "^8.7.0",
     "prettier": "^2.8.4"
