@@ -1,19 +1,10 @@
-import Head from 'next/head';
-import * as React from 'react';
-import { dAppName } from '../config';
-import { routeNames } from '../routes';
-import Link from 'next/link';
+import {dAppName} from "@/config";
+import Link from "next/link";
+import {routeNames} from "@/routes";
+import * as React from "react";
 
-export default function Index() {
+export default function Home() {
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name='description' content='Generated by create next app' />
-        <meta name='viewport' content='width=device-width, initial-scale=1' />
-        <link rel='icon' href='/favicon.ico' />
-      </Head>
-
       <main className='mt-5'>
         <div className='d-flex flex-fill align-items-center container'>
           <div className='row w-100'>
@@ -30,9 +21,9 @@ export default function Index() {
                   </p>
 
                   <Link
-                    href={routeNames.unlock}
-                    className='btn btn-primary mt-3 text-white'
-                    data-testid='loginBtn'
+                      href={routeNames.unlock}
+                      className='btn btn-primary mt-3 text-white'
+                      data-testid='loginBtn'
                   >
                     Login
                   </Link>
@@ -42,6 +33,5 @@ export default function Index() {
           </div>
         </div>
       </main>
-    </>
-  );
+  )
 }
