@@ -1,8 +1,10 @@
+'use client';
+
 import React, { PropsWithChildren } from 'react';
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import Link from 'next/link';
-import { routeNames } from '../../../routes';
+import { routeNames } from '@/routes';
 
 export const StatisticsLayout = ({ children }: PropsWithChildren) => {
   return (
