@@ -1,12 +1,14 @@
+"use client"
+
 import React, { useEffect, useState } from 'react';
 import { faTriangleExclamation } from '@fortawesome/free-solid-svg-icons';
 import axios from 'axios';
 import Link from 'next/link';
 
 import { StatisticsLayout } from './components/StatisticsLayout';
-import { routeNames } from '../../routes';
-import { Loader, PageState } from '../../components';
-import { TOOLS_API_URL } from '../../config';
+import { routeNames } from '@/routes';
+import { Loader, PageState } from '@/components';
+import { TOOLS_API_URL } from '@/config';
 import { useGetLoginInfo } from '@multiversx/sdk-dapp/hooks';
 
 interface StatisticsType {
