@@ -1,8 +1,9 @@
+'use client';
+
 import React from 'react';
-import { walletConnectV2ProjectId } from '../../config';
-import { routeNames } from '../../routes';
-import { AuthRedirectWrapper } from '../../components/AuthRedirectWrapper';
-import Head from 'next/head';
+import { walletConnectV2ProjectId } from '@/config';
+import { routeNames } from '@/routes';
+import { AuthRedirectWrapper } from '@/components/AuthRedirectWrapper';
 import dynamic from 'next/dynamic';
 
 const ExtensionLoginButton = dynamic(
@@ -50,12 +51,6 @@ const UnlockPage = () => {
 
   return (
     <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name='description' content='Generated by create next app' />
-        <meta name='viewport' content='width=device-width, initial-scale=1' />
-        <link rel='icon' href='/favicon.ico' />
-      </Head>
       <main className='mt-5'>
         <div className='home d-flex flex-fill align-items-center'>
           <div className='m-auto' data-testid='unlockPage'>
