@@ -4,8 +4,8 @@ import {
   SmartContract,
   Address
 } from '@multiversx/sdk-core/out';
-import { contractAddress } from '../../../../config';
-import json from '../../../../ping-pong.abi.json';
+import { contractAddress } from '@/config';
+import json from '@/abi/ping-pong.abi.json';
 
 const abiRegistry = AbiRegistry.create(json);
 const abi = new SmartContractAbi(abiRegistry);
