@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useGetAccountInfo } from '@multiversx/sdk-dapp/hooks';
-import { contractAddress } from '../../config';
+import { contractAddress } from '@/config';
 import dynamic from 'next/dynamic';
 
 const FormatAmount = dynamic(
